Add use_pos for calculating xmat
So now I can skip reading xmat from shirley_xas.
If use_pos = True, the code will look for the *.pos file for the
corresponding pseudopotential in the pseudo library, and use the
position matrix elements in that file to calculate the transition
matrix elements for all bands in the system (emtpy or occupied):

	xmat:	< nk | r_i | h_c >

xmat is an array of nbnd x ncp x nxyz.
n in range(nbnd)
c in range(ncp)
i in range(nxyz)

This is particularly useful for:
(1) cases where there are more than one excited atom
(2) non K edges situations where there are multiple core levels,
though rehybridization of the core levels under SO coupling still
needs to be coded.

This part needs to be further tested for more examples.