First version for li metals
Although it may be only for lithium metals for now, it can be
extended to a more general situation. In fact, I think it can be
used in any situations. Major changes:

(1) Add DFS for searching for determinants, using major matrix elements
(2) Directly calculate determinants
(3) Group configurations with similar energies so as to reduce #
configurations