Critical bug: ranking of zeta matrix elements
The matrix elements of zeta should be ranked according to their
absolute values, not the real parts. If one compare two complex numbers
c1 and c2 like this: c1 < c2, then only the real parts will be used
for comparison in python. Actually, abs(c1) < abs(c2) is the correct way.

This will change the spectrum to a significnat degree because some
import matrix elements would be missing if using the real parts only
(consider the situation where the real part is small whereas the imaginary
part is big. )