Critical bug fixed in dfs for xps
The matrix element for f(1) of xps would be just the matrix elements
of the zeta matrix. The previous calculation method as implemented in
the dfs_cv would ignore the matrix elements on the last column of the
zeta matrix. So that will result in a smaller xps and hence xas.